rename lvs -> csilvm
