+++ pkg/csilvm/client.go
@@ -1,4 +1,4 @@
-package lvs
+package csilvm
 
 import (
 	"github.com/container-storage-interface/spec/lib/go/csi"
