+++ pkg/csilvm/mount.go
@@ -1,4 +1,4 @@
-package lvs
+package csilvm
 
 import (
 	"io/ioutil"
