+++ pkg/csilvm/validate_test.go
@@ -1,4 +1,4 @@
-package lvs
+package csilvm
 
 import (
 	"context"
