+++ pkg/lvm/scan.go
